+++ frontend/src/metabase/admin/databases/components/DatabaseConnectionInfoSectionContent/DatabaseConnectionInfoSectionContent.module.css
@@ -1,0 +1,3 @@
+.actionButton[data-action-status="idle"] {
+  background-color: var(--mb-color-bg-white);
+}
