+++ frontend/src/metabase/admin/databases/components/DatabaseConnectionInfoSectionContent/index.tsx
@@ -1,0 +1,1 @@
+export * from "./DatabaseConnectionInfoSectionContent";
