+++ frontend/src/metabase/admin/databases/components/DatabaseInfoSection.tsx
@@ -4,10 +4,12 @@ export const DatabaseInfoSection = ({
   name,
   description,
   children,
+  condensed = false,
 }: {
   name: string;
   description: string;
   children: React.ReactNode;
+  condensed?: boolean;
 }) => (
   <Flex gap="5.5rem" mb="5.5rem">
     <Flex gap="lg" w="100%">
@@ -20,11 +22,11 @@ export const DatabaseInfoSection = ({
       <Box w="100%">
         <Card
           withBorder
-          shadow="none"
           bg="bg-light"
+          shadow="none"
           style={{ flexGrow: 0 }}
-          px="1.375rem"
-          py="1.875rem"
+          px="1.5rem"
+          py={condensed ? "1.125rem" : "1.875rem"}
         >
           {children}
         </Card>
@@ -32,3 +34,17 @@ export const DatabaseInfoSection = ({
     </Flex>
   </Flex>
 );
+
+// NOTE: not using Card.Section since it won't work if
+// it's rendered within a React.Fragment: https://mantine.dev/core/card/#cardsection
+export const DatabaseInfoSectionDivider = () => {
+  return (
+    <Box
+      h="1px"
+      w="calc(100% + 3rem)"
+      ml="-1.5rem"
+      my="1rem"
+      bg="var(--mb-color-border)"
+    />
+  );
+};
