+++ frontend/src/metabase/admin/databases/components/DatabaseModelFeaturesSectionContent/ModelActionsSection/index.ts
