+++ frontend/src/metabase/admin/databases/components/DatabaseModelFeaturesSectionContent/ModelCachingControl/index.ts
