+++ frontend/src/metabase/admin/databases/components/DatabaseModelFeaturesSectionContent/index.tsx
@@ -1,0 +1,1 @@
+export * from "./DatabaseModelFeaturesSectionContent";
