--- frontend/src/metabase/admin/databases/containers/DatabaseEditApp.styled.tsx
@@ -1,10 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-import styled from "@emotion/styled";
-
-import { DatabaseHelpCard } from "metabase/databases/components/DatabaseHelpCard";
-
-export const DatabaseEditHelp = styled(DatabaseHelpCard)`
-  max-width: 21rem;
-  margin-top: 1.25rem;
-  margin-left: 6.5rem;
-`;
