+++ frontend/src/metabase/admin/databases/containers/DatabaseEditApp.tsx
@@ -1,7 +1,5 @@
 import type { Location, LocationDescriptor } from "history";
-import { updateIn } from "icepick";
 import type { ComponentType } from "react";
-import { useState } from "react";
 import type { Route } from "react-router";
 import { push } from "react-router-redux";
 import { useMount } from "react-use";
@@ -9,19 +7,15 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
-import Sidebar from "metabase/admin/databases/components/DatabaseEditApp/Sidebar/Sidebar";
 import Breadcrumbs from "metabase/components/Breadcrumbs";
 import { GenericError } from "metabase/components/ErrorPages";
-import { LeaveConfirmationModal } from "metabase/components/LeaveConfirmationModal";
-import { LoadingAndErrorWrapper } from "metabase/components/LoadingAndErrorWrapper";
 import CS from "metabase/css/core/index.css";
-import { DatabaseForm } from "metabase/databases/components/DatabaseForm";
 import title from "metabase/hoc/Title";
-import { useCallbackEffect } from "metabase/hooks/use-callback-effect";
+import { useModal } from "metabase/hooks/use-modal";
 import { connect } from "metabase/lib/redux";
 import { getSetting } from "metabase/selectors/settings";
 import { getUserIsAdmin } from "metabase/selectors/user";
-import { Box, Divider, Flex } from "metabase/ui";
+import { Box, Divider, Modal } from "metabase/ui";
 import Database from "metabase-lib/v1/metadata/Database";
 import type {
   DatabaseData,
@@ -32,19 +26,21 @@ import type { State } from "metabase-types/store";
 
 import { DatabaseConnectionInfoSectionContent } from "../components/DatabaseConnectionInfoSectionContent";
 import { DatabaseDangerZoneSectionContent } from "../components/DatabaseDangerZoneSectionContent";
+import {
+  DatabaseEditConnectionForm,
+  type DatabaseEditErrorType,
+} from "../components/DatabaseEditConnectionForm";
 import { DatabaseInfoSection } from "../components/DatabaseInfoSection";
+import { DatabaseModelFeaturesSectionContent } from "../components/DatabaseModelFeaturesSectionContent";
 import { ExistingDatabaseHeader } from "../components/ExistingDatabaseHeader";
 import {
   dismissSyncSpinner,
   initializeDatabase,
   reset,
-  saveDatabase,
   selectEngine,
   updateDatabase,
 } from "../database";
 import { getEditingDatabase, getInitializeError } from "../selectors";
-
-import { DatabaseEditHelp } from "./DatabaseEditApp.styled";
 
 interface DatabaseEditAppProps {
   database?: Database;
@@ -52,7 +48,6 @@ interface DatabaseEditAppProps {
   reset: () => void;
   initializeDatabase: (databaseId: DatabaseId) => void;
   dismissSyncSpinner: (databaseId: DatabaseId) => Promise<void>;
-  saveDatabase: (database: DatabaseData) => Database;
   updateDatabase: (
     database: { id: DatabaseId } & Partial<DatabaseType>,
   ) => Promise<void>;
@@ -60,6 +55,7 @@ interface DatabaseEditAppProps {
   location: Location;
   isAdmin: boolean;
   isModelPersistenceEnabled: boolean;
+  // TODO: remove
   initializeError?: DatabaseEditErrorType;
   route: Route;
   onChangeLocation: (location: LocationDescriptor) => void;
@@ -79,48 +75,29 @@ const mapStateToProps = (state: State) => {
 const mapDispatchToProps = {
   reset,
   initializeDatabase,
-  saveDatabase,
   updateDatabase,
   dismissSyncSpinner,
   selectEngine,
   onChangeLocation: push,
 };
 
-type DatabaseEditErrorType = {
-  data: {
-    message: string;
-    errors: { [key: string]: string };
-  };
-  statusText: string;
-  message: string;
-};
-
-function DatabaseEditApp(props: DatabaseEditAppProps) {
-  const {
-    database,
-    updateDatabase,
-    initializeError,
-    dismissSyncSpinner,
-    isAdmin,
-    isModelPersistenceEnabled,
-    reset,
-    initializeDatabase,
-    params,
-    saveDatabase,
-    route,
-    onChangeLocation,
-  } = props;
-
+function DatabaseEditAppInner({
+  database,
+  updateDatabase,
+  initializeError,
+  dismissSyncSpinner,
+  isAdmin,
+  isModelPersistenceEnabled,
+  reset,
+  initializeDatabase,
+  params,
+  route,
+  onChangeLocation,
+}: DatabaseEditAppProps) {
   const editingExistingDatabase = database?.id != null;
   const addingNewDatabase = !editingExistingDatabase;
 
-  const [isDirty, setIsDirty] = useState(false);
-
-  /**
-   * Navigation is scheduled so that LeaveConfirmationModal's isEnabled
-   * prop has a chance to re-compute on re-render
-   */
-  const [isCallbackScheduled, scheduleCallback] = useCallbackEffect();
+  const dbDetailsModal = useModal(false);
 
   useMount(async () => {
     await reset();
@@ -131,124 +108,81 @@ function DatabaseEditApp(props: DatabaseEditAppProps) {
     [t`Databases`, "/admin/databases"],
     [addingNewDatabase ? t`Add Database` : database.name],
   ];
-  const handleSubmit = async (database: DatabaseData) => {
-    try {
-      const savedDB = await saveDatabase(database);
-      if (addingNewDatabase) {
-        scheduleCallback(() => {
-          onChangeLocation(
-            `/admin/databases?created=true&createdDbId=${savedDB.id}`,
-          );
-        });
-      }
-    } catch (error) {
-      throw getSubmitError(error as DatabaseEditErrorType);
-    }
-  };
-
-  const autofocusFieldName = window.location.hash.slice(1);
-
   // TODO: handle this on a new page
   if (addingNewDatabase) {
     return <div>TODO: b0rked this on purpose for the time being</div>;
   }
 
+  // TODO: && a check for if any section in the models section should be shown
+  const shouldShowModelFeaturesSection = !database.is_attached_dwh;
+  const shouldShowDangerZone = !database.is_attached_dwh;
+
   return (
-    <Box w="100%" maw="64.25rem" mx="auto" px="2rem">
-      <Breadcrumbs className={CS.py4} crumbs={crumbs} />
+    <ErrorBoundary errorComponent={GenericError as ComponentType}>
+      <Box w="100%" maw="64.25rem" mx="auto" px="2rem">
+        <Breadcrumbs className={CS.py4} crumbs={crumbs} />
 
-      <ExistingDatabaseHeader database={database} />
+        <ExistingDatabaseHeader database={database} />
 
-      <Divider mb="3.25rem" />
+        <Divider mb="3.25rem" />
 
-      <DatabaseInfoSection
-        name={t`Connection and sync`}
-        description={t`Manage details about the database connection and when Metabase ingests new data.`}
-      >
-        <DatabaseConnectionInfoSectionContent database={database} />
-      </DatabaseInfoSection>
-
-      <DatabaseInfoSection
-        name={t`Model features`}
-        description={t`A Jeff B diddy`}
-      >
-        <Flex align="center" justify="space-between">
-          TODO
-        </Flex>
-      </DatabaseInfoSection>
-
-      <DatabaseInfoSection
-        name={t`Danger zone`}
-        description={t`Remove this database and other destructive actions`}
-      >
-        <DatabaseDangerZoneSectionContent
-          isAdmin={isAdmin}
-          database={database}
-        />
-      </DatabaseInfoSection>
-
-      <Flex mb="md">
-        <ErrorBoundary errorComponent={GenericError as ComponentType}>
-          <div>
-            <div className={CS.pt0}>
-              <LoadingAndErrorWrapper
-                loading={!database}
-                error={initializeError}
-              >
-                {editingExistingDatabase && database.is_attached_dwh ? (
-                  <div>{t`This database cannot be modified.`}</div>
-                ) : (
-                  <Flex>
-                    <Box w="38.5rem">
-                      <DatabaseForm
-                        initialValues={database}
-                        isAdvanced
-                        onSubmit={handleSubmit}
-                        setIsDirty={setIsDirty}
-                        autofocusFieldName={autofocusFieldName}
-                      />
-                    </Box>
-                    <div>{addingNewDatabase && <DatabaseEditHelp />}</div>
-                  </Flex>
-                )}
-              </LoadingAndErrorWrapper>
-            </div>
-          </div>
-        </ErrorBoundary>
-
-        {editingExistingDatabase && !database.is_attached_dwh && (
-          <Sidebar
+        <DatabaseInfoSection
+          name={t`Connection and sync`}
+          description={t`Manage details about the database connection and when Metabase ingests new data.`}
+          condensed
+        >
+          <DatabaseConnectionInfoSectionContent
             database={database}
-            isAdmin={isAdmin}
-            isModelPersistenceEnabled={isModelPersistenceEnabled}
-            updateDatabase={updateDatabase}
             dismissSyncSpinner={dismissSyncSpinner}
+            openDbDetailsModal={dbDetailsModal.open}
           />
-        )}
-      </Flex>
+        </DatabaseInfoSection>
 
-      <LeaveConfirmationModal
-        isEnabled={isDirty && !isCallbackScheduled}
-        route={route}
-      />
-    </Box>
+        {shouldShowModelFeaturesSection && (
+          <DatabaseInfoSection
+            name={t`Model features`}
+            description={t`Choose whether to enable features related to Metabase models. These will often require a write connection.`}
+          >
+            <DatabaseModelFeaturesSectionContent
+              database={database}
+              isModelPersistenceEnabled={isModelPersistenceEnabled}
+              updateDatabase={updateDatabase}
+            />
+          </DatabaseInfoSection>
+        )}
+
+        {shouldShowDangerZone && (
+          <DatabaseInfoSection
+            name={t`Danger zone`}
+            description={t`Remove this database and other destructive actions`}
+          >
+            <DatabaseDangerZoneSectionContent
+              isAdmin={isAdmin}
+              database={database}
+            />
+          </DatabaseInfoSection>
+        )}
+
+        <Modal
+          title={t`Edit connection details`}
+          opened={dbDetailsModal.opened}
+          onClose={dbDetailsModal.close}
+        >
+          <DatabaseEditConnectionForm
+            database={database}
+            initializeError={initializeError}
+            onChangeLocation={onChangeLocation}
+            route={route}
+          />
+        </Modal>
+      </Box>
+    </ErrorBoundary>
   );
 }
 
-const getSubmitError = (error: DatabaseEditErrorType) => {
-  if (_.isObject(error?.data?.errors)) {
-    return updateIn(error, ["data", "errors"], errors => ({
-      details: errors,
-    }));
-  }
-
-  return error;
-};
-
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export default _.compose(
+export const DatabaseEditApp = _.compose(
   connect(mapStateToProps, mapDispatchToProps),
   title(
     ({ database }: { database: DatabaseData }) => database && database.name,
   ),
-)(DatabaseEditApp);
+)(DatabaseEditAppInner);
