+++ frontend/src/metabase/components/ActionButton/ActionButton.jsx
@@ -122,10 +122,13 @@ export default class ActionButton extends Component {
     const { active, result } = this.state;
     const isActionDisabled = active || result === "success";
 
+    const actionStatus = active ? "pending" : (result ?? "idle");
+
     return (
       <Button
         ref={innerRef}
         {...props}
+        data-action-status={actionStatus}
         className={
           forceActiveStyle
             ? ButtonsS.Button
